Have a single file for every copy method
Moved every way of copying to its own file. Went for a new to distinguish between the ways of copying by passing a copy function as parameter; We now either copy simple, by tech or do the atomic (per file) way
